requirements.txt
numpy
pandas
opencv-python
scikit-learn
tqdm
xgboost
joblib
torch
torchvision
torchaudio
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
Pillow
